Add unit tester
