+++ logs/README.md
@@ -1,0 +1,2 @@
+# Logs
+Run logs and debug output.
