+++ lynx/__init__.py
@@ -1,0 +1,10 @@
+"""Lynx video upscaling package."""
+
+__all__ = [
+    "download",
+    "models",
+    "upscale",
+    "encode",
+    "processor",
+    "gui",
+]
