+++ lynx/__main__.py
@@ -1,0 +1,4 @@
+from .gui import main
+
+if __name__ == "__main__":
+    main()
