+++ outputs/README.md
@@ -1,0 +1,2 @@
+# Outputs
+Encoded MP4 files are placed here.
