+++ weights/README.md
@@ -1,0 +1,2 @@
+# Weights
+Store Real-ESRGAN model weights here.
