+++ work/README.md
@@ -1,0 +1,2 @@
+# Work Folder
+Temporary downloads and processing files.
